feat: harden order and stock handling
Fix the correctness gaps around stock accounting, concurrent orders and money
precision, and lock the behaviour down with tests.

- Product: add @Version optimistic locking so two concurrent orders cannot
  oversell the same stock; price becomes BigDecimal mapped to NUMERIC(19,2)
- OrderItem: unitPrice becomes BigDecimal (NUMERIC(19,2)), snapshotting the
  price at order time
- OrderService.create(): validate the whole request first and only then mutate
  stock, so a rejected request never leaves stock partially decremented;
  reject a duplicate productId with a descriptive 400 instead of an opaque 409
- OrderService.deleteById(): give the reserved quantities back, previously
  deleting an order leaked stock permanently
- ProductService.deleteById(): report a product still referenced by orders as
  a specific 409 via OrderItemRepository.countByProductId
- OrderRepository: @EntityGraph on findAll/findById to remove the N+1 (7
  queries down to 1 for 3 orders / 6 items)
- GlobalExceptionHandler: add ConflictException and map
  OptimisticLockingFailureException / DataIntegrityViolationException to 409
- ProductRequest: stock is now required, price is a BigDecimal limited to
  2 decimals via @Digits

Tests (27 total) run against an isolated in-memory H2 database so they never
touch the on-disk demo1db.mv.db. The test classes deliberately omit
@Transactional so each service call really commits and rollback behaviour is
verified rather than masked.

+++ src/main/java/com/example/demo1/dto/OrderItemResponse.java
@@ -5,31 +5,22 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.math.BigDecimal;
 
-/**
- * Response representation of a single order line item.
- */
 @Schema(description = "Order line item returned to the client")
 public record OrderItemResponse(
 
-        @Schema(description = "ID of the product", example = "1")
-        Long productId,
+        @Schema(description = "ID of the product", example = "1") Long productId,
 
-        @Schema(description = "Product name", example = "Laptop")
-        String productName,
+        @Schema(description = "Product name", example = "Laptop") String productName,
 
-        @Schema(description = "Quantity ordered", example = "2")
-        Integer quantity,
+        @Schema(description = "Quantity ordered", example = "2") Integer quantity,
 
-        @Schema(description = "Unit price at order time", example = "15000000")
-        BigDecimal unitPrice
-) {
+        @Schema(description = "Unit price at order time", example = "15000000") BigDecimal unitPrice) {
 
-    /** Maps an entity to a response DTO, avoiding floating-point scientific notation. */
     public static OrderItemResponse from(OrderItem item) {
         return new OrderItemResponse(
                 item.getProduct().getId(),
                 item.getProduct().getName(),
                 item.getQuantity(),
-                BigDecimal.valueOf(item.getUnitPrice()));
+                item.getUnitPrice());
     }
 }
